finished lesson 11, added gitignore to ignore ./idea files generated by RubyMine
