+++ .gitignore
@@ -1,0 +1,2 @@
+# Ignore .idea files produced by RubyMine
+.idea/
